server/requirements.txt
Flask==1.1.1
numpy==1.17.4
Pillow==6.2.1
PyWavelets==1.1.1